// frontend/src/shared/api/models/IMovie.ts
export interface IMovie {
    id: number; 
    title: string; 
    movie_description: string; 
    release_date: string; 
    budget: number; 
    country: string;
    production_companies: string;
    actors: string; 
    director: string;
    writers: string; 
    ratings: number; 
    duration: number;
    genre: string; 
    poster_image: string; 
}